+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@reside-ic/odin",
-    "version": "0.0.2",
+    "version": "0.0.3",
     "description": "odin js/ts support",
     "main": "lib/index.js",
     "types": "lib/index.d.ts",
