+++ src/fit.ts
@@ -1,28 +1,32 @@
-import {Simplex} from "dfoptim";
-
 import { base } from "./base";
 import type { OdinModelConstructable, Solution } from "./model";
 import {interpolatedSolution, partialInterpolatedSolution, runModel} from "./model";
 import type {UserType} from "./user";
 
+/** Interface for data to fit an odin model to; every data set has two
+ *  series, even if they are derived from some larger data set.
+ */
 export interface FitData {
+    /** Array of time values; must be increasing */
     time: number[];
+    /** Observed data to fit to; must be the same length as `time` */
     value: number[];
 }
 
+/** Interface to control parameters of a fit. A model will have some
+ *  number of parameters but only some change. We also need to provide
+ *  starting points for the fit.
+ */
 export interface FitPars {
+    /** Base parameters; used as starting points for those that change
+     * (within `vary`) and fixed values for those that don't change
+     */
     base: UserType;
+    /** Array of names of parameters that should be optimised during
+     * the fitting process. All values in `vary` must be present as
+     * keys in `base`.
+     */
     vary: string[];
-}
-
-export function startFit(Model: OdinModelConstructable, data: FitData,
-                         pars: FitPars, modelledSeries: string,
-                         controlODE: any, controlFit: any) {
-    const target = fitTarget(Model, data, pars, modelledSeries, controlODE);
-    // TODO: require that we have starting points here (i.e., that
-    // everything variable is in fact a number)
-    const start = pars.vary.map((nm: string) => pars.base.get(nm) as number);
-    return new Simplex(target, start, controlFit);
 }
 
 export function fitTarget(Model: OdinModelConstructable,
@@ -31,32 +35,32 @@ export function fitTarget(Model: OdinModelConstructable,
                           control: any) {
     const tStart = data.time[0];
     const tEnd = data.time[data.time.length - 1];
-    // TODO: it would be preferable if this error handling moved into
-    // dfoptim, as it will be fairly common; could make it an option
-    // for the control, which would be much nicer and easier to test.
     return (theta: number[]) => {
         const p = updatePars(pars, theta);
         const model = new Model(base, p, "error");
         const y0 = null;
-        const solution =
-            runModel(model, y0, tStart, tEnd, control).solution;
-
+        const solution = runModel(model, y0, tStart, tEnd, control).solution;
         const names = model.names();
         const idxModel = names.indexOf(modelledSeries);
         const yModel = solution(data.time).map((y) => y[idxModel]);
-        const value = sumOfSquares(data.value, yModel);
 
-        // We return a solution-type function just for the modelled
-        // series as that's probably what the caller will
-        // want. However, the full solution will be available too,
-        // alongside the names.
-        const solutionFit = partialInterpolatedSolution(
-            solution, modelledSeries, idxModel, tStart, tEnd);
-        const solutionAll = interpolatedSolution(
-            solution, names, tStart, tEnd);
-
-        return { data: {names, pars: p, solutionAll, solutionFit},
-                 value };
+        // Unfortunately, TypeDoc does not copy all doc comments here
+        // over, so I've not documented them here.
+        return {
+            /** Additional data alongside the goodness of fit, see above */
+            data: {
+                names,
+                pars: p,
+                solutionAll: interpolatedSolution(
+                    solution, names, tStart, tEnd),
+                solutionFit: partialInterpolatedSolution(
+                    solution, modelledSeries, idxModel, tStart, tEnd),
+            },
+            /** Goodness of fit, the sum-of-squared differences
+             * between the observed data and the modelled series
+             */
+            value: sumOfSquares(data.value, yModel),
+        };
     };
 }
