+++ test/wodin.test.ts
@@ -125,13 +125,30 @@ describe("can set user", () => {
     })
 });
 
-describe("can fit a model", () => {
-    it("fits", () => {
+describe("can fit a simple line", () => {
+    it("Can fit a simple model", () => {
         const time = [0, 1, 2, 3, 4, 5, 6];
         const data = {time, value: time.map((t: number) => 1 + t * 4)}
-        const pars = {base: new Map<string, number>([["a", 0.5]]), vary: ["a"]};
-        const opt = wodinFit(models.User, data, pars, "x", {}, {});
+        const pars = {base: new Map<string, number>([["a", 0.5]]),
+                      vary: ["a"]};
+        const modelledSeries = "x";
+        const controlODE = {};
+        const controlFit = {};
+
+        const opt = wodinFit(models.User, data, pars, modelledSeries,
+                             controlODE, controlFit);
         const res = opt.run(100);
         expect(res.converged).toBe(true);
+        expect(res.location[0]).toBeCloseTo(4);
+        expect(res.value).toBeCloseTo(0);
+        expect(res.data.pars.get("a")).toEqual(res.location[0]);
+
+        const yFit = res.data.solutionFit(0, 6, 7);
+        expect(yFit.name).toEqual("x");
+        expect(yFit.x).toEqual(time);
+        expect(approxEqualArray(yFit.y, data.value)).toBe(true);
+
+        const yFull = res.data.solutionAll(0, 6, 7);
+        expect(yFull).toEqual([yFit]);
     });
 });
